requirements.txt
alembic==1.4.2
aniso8601==8.0.0
attrs==19.3.0
click==7.1.2
Flask==1.1.2
Flask-Migrate==2.5.3
flask-restplus==0.13.0
Flask-SQLAlchemy==2.4.3
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.3.0
packaging==20.4
pkg-resources==0.0.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.7
pyrsistent==0.16.0
pytest==5.4.3
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.17
wcwidth==0.2.4
Werkzeug==0.16.1
flask_login
